// src/app/interfaces/user.interface.ts
export interface User {
    id: number;
    name: string;
    username?: string;
    email: string;
    address?: {
      street?: string;
      suite?: string;
      city?: string;
      zipcode?: string;
      geo?: {
        lat?: string;
        lng?: string;
      };
    };
    phone?: string;
    website: string;
    company: {
      name: string;
    }
  }
  export interface IUser {
    id: number;
    name: string;
    email: string;
    phone?: string;
    website: string;
    company: {
     name: string
    }
  }